More whitespace between cell content and cell border
The row height is 4 pix higher than the actual font height --> so two pix on top and below the actual text within the cell.

The cell renderer leaves 6pix space on the left and right side of the cell content.

Both changes lead to much better readability of table contents.