Alternating color for tables, inverted cursor
All tables have alternating row colors (pastel blue, white).

The selection color of the current cell is a dark blue and white text.

When a row is selected, text color is always white - so the disabled text colors and highlighted text colors (red) will only be applied on rows which are not selected.